dwarf: Fix the incorrect enum display problem
The enum value can contain a negative value, but it's incorrectly
displayed in the current implementation.

Let's say there is a simple program as follows.

  $ cat enum.c
  enum et {
          MINUS_ONE = -1,
          ZERO,
          ONE,
  };

  void foo(enum et e) {}

  int main() {
          foo(MINUS_ONE);
          return 0;
  }

The tracing output with auto-args is as follows.

  $ gcc -pg -g enum.c
  $ uftrace -a a.out
  # DURATION     TID     FUNCTION
              [ 56519] | main() {
     4.446 us [ 56519] |   foo(ONE|ZERO|MINUS_ONE+0xffffffff);
     4.957 us [ 56519] | } = 0; /* main */

The size of enum by default is 4, but uftrace currently expects the
sizes is to be sizeof(long) so 8.

So the negative value -1 in enum type is interpreted as 8 bytes so the
it reads more bytes than stored, but it can be interpreted incorrectly.

This patch assumes the size of enum is 4 so type casted as int type and
it fixes the problem.

  $ uftrace -a a.out
  # DURATION     TID     FUNCTION
              [ 22103] | main() {
     9.033 us [ 22103] |   foo(MINUS_ONE);
    11.284 us [ 22103] | } = 0; /* main */

Fixed: #1154

Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>