script/python: Work around build failure due to _Py_Dealloc
Starting from python 3.8, it seems _Py_Dealloc() function is required
by Py_DECREF macro which is widely used here and there.  Normally it
should be ok but as we load the library with dlopen, we should not
call it directly.

Thus, let's copy the marco (in a minimal fashion), and call it with
dynamic symbol.  This fixes the following error:

  /usr/bin/ld: /home/namhyung/project/uftrace/libmcount/script-python.op: in function `_Py_DECREF':
  /usr/include/python3.8/object.h:478: undefined reference to `_Py_Dealloc'
  /usr/bin/ld: /usr/include/python3.8/object.h:478: undefined reference to `_Py_Dealloc'
  /usr/bin/ld: /usr/include/python3.8/object.h:478: undefined reference to `_Py_Dealloc'
  /usr/bin/ld: /usr/include/python3.8/object.h:478: undefined reference to `_Py_Dealloc'
  /usr/bin/ld: /usr/include/python3.8/object.h:478: undefined reference to `_Py_Dealloc'
  /usr/bin/ld: /usr/include/python3.8/object.h:478: more undefined references to `_Py_Dealloc' follow
  collect2: error: ld returned 1 exit status

Fixed: #1190

Signed-off-by: Namhyung Kim <namhyung@gmail.com>