script: Add version info to uftrace_script_info
This patch adds script version info to uftrace_script_info and the
script version can be changed when the layout of the script related
struct types are changed.

The script writer has a reponsibility to check the version. The API
version of the script is SCRIPT_API_VERSION and the API version of
uftrace engine is sc_info->api_version.

Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>