build: Add initial files for meson build system
This build system goal is to replace the existing
one based on makefiles. It is both faster and
easier to maintain that the existing one.

The following lists examples on how to configure
and the build the project. Both `meson` and
`ninja` commands are required.

To use it, first create a build directory:

 $ mkdir build && cd build

To configure build, execute one of these:

 $ meson ..          # automatic configuration
 $ CC=clang meson .. # change compiler
 $ meson .. -Dwith-libpython=disabled

To list options, execute:

 $ meson configure

To reconfigure an existing build, execute:

 $ meson --reconfigure -Dwith-libpython=disabled

To build uftrace, execute:

 $ ninja

To install uftrace, execute one of these:

 # meson install             # system directories
 $ DESTDIR=pkg meson install # building a package

To run tests, execute:

 $ ninja unittest # run unit tests
 $ ninja runtest  # run uftrace tests
 $ ninja tests    # run unit and uftrace tests

NOTE: Cross-compilation is rudimentary and not
well-tested, but it is built-in in Meson [1]. For
additional informations concerning Meson, see [2]
or [3].

[1] https://mesonbuild.com/Cross-compilation.html
[2] https://mesonbuild.com/howtox.html
[3] https://mesonbuild.com/Quick-guide.html

Signed-off-by: Gabriel-Andrew Pollo-Guilbert <gabrielpolloguilbert@gmail.com>