dynamic: local variable min_size moved to file scope
The min_size means the minimum size of the function to be
instrumented.  It's parsed from an environment variable when try to
update the code at the start.

But later, it's also needed for dlopen functions.  However, if process
changed/removed enviroments it might not see the original value.  So
it'd better keeping it as the original value using a static variable.

Signed-off-by: Hanbum Park <kese111@gmail.com>
Signed-off-by: Namhyung Kim <namhyung@gmail.com>