demangle: Fix compiler warning on dd_append_len()
Some version of GCC complained that strncpy() with the exact string
length is dangerous since it makes the final string not terminated
properly.  Although our code is safe since it manually terminates the
string, it's better to inform the compiler explicitly.

We can do it by passing size + 1 as the length then strncpy() will
terminate the string for us.  While it's true for normal cases, for
demangling it usually passes a part of a string to extract that so we
cannot simply rely on it.  So we need to keep the manual termination.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>