report: Fix memory use-after-free in diff mode
The uftrace report --diff deals with two data files so it needs to
call close_data_file() twice.  But the modules are global resources
and it's released from the first call and then second will refer
dangling pointers.

Fix it by splitting close_data_file() to control module unload.  This
handle a part of problems in #1213.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>