dynamic/x86: Skip *.cold functions
The cold functions are splitted from the original function because of
some optimizations.  So they likely have jumps to the middle of the
original function.  Actually they are detected by the badsym logic but
it ends up with preventing the original functions too.

If the original function has no other reason to fail, we can use it by
simply skipping the matching cold function.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>