plthook: Restore return address properly on AArch64
On ARM and AArch64, return address is saved in the LR register not on
the stack.  So at the time plt_hooker is called, it cannot know the
location of return address on the stack.  Actually it's saved when the
actual (library) function is called.  So ARCH_CAN_RESTORE_PLTHOOK
macro is used to check the condition.

Note that it's ok for uftrace to trace PLT functions since the return
address is changed during the plt_hooker so it'll call plthook_return
after running the function.

But in case of C++ exception, all the return addresses we hooked
should be restored in order to internal stack unwinding works
correctly.  So we should scan the stack to find location where the
return address of a PLT function is saved.

Fixed: #1020

Signed-off-by: Namhyung Kim <namhyung@gmail.com>