mcount: dynamic: Split mcount_dynamic_update initialisation
In order to dynamically patch and unpatch functions, the
mcount_dynamic_update function will be called each time.
That said, it currently is assumed that this function is
only called once at the initialisation of libmcount.

This commit splits parts of the function that must only
be called once into its own initialisation function.

The loading of the dynamic module informations has also
been updated to be called multiple time. The first time
it is called, it loads either the main binary or all
modules. The next time it's called, it only load other
modules if necessary.

Signed-off-by: Gabriel-Andrew Pollo-Guilbert <gabrielpolloguilbert@gmail.com>