tests: Fix 051 return test when using clang
When testing t051_return.py with aarch64 clang, it fails for two
reasons.  The first reason is that aarch64 sometimes add memcpy for
struct construction.

  t051_return: diff result of -pg -O0
  --- expect      2021-04-23 18:20:58.426790428 +0000
  +++ result      2021-04-23 18:20:58.426790428 +0000
  @@ -3,4 +3,4 @@
        memset();
  -     memcpy();
      } /* return_large */
  +   memcpy();
      return_small();

The second reason is because clang sometimes ignores -fno-inline and
does constant folding for very simple functions.  To suppress this, we
need to add __attribute__((optnone)), which is a clang specific
attribute.

  t051_return: diff result of -pg -O1
  --- expect      2021-04-23 18:21:54.182494978 +0000
  +++ result      2021-04-23 18:21:54.182494978 +0000
  @@ -3,6 +3,3 @@
        memset();
  -     memcpy();
      } /* return_large */
  -   return_small();
  -   return_long_double();
    } /* main */

This patch fixes those two problems together.

Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>