tests: Fix 136 dynamic test in other arch
Since 136 dynamic test is only for x86_64, it should be disabled in
other architectures.

To do this, this patch adds TestBase.get_machine() and use it to skip
the test in other architectures.

We also need to skip this test for clang because -mfentry -mnop-mcount
options are supported only in gcc.

Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>