mcount: Start daemon in libmcount
This patch introduces libmcount version containing a daemon, waiting for
incoming commands.

The daemon is started as a new thread when libmcount is set up. It is
listening to a socket at /var/run/user/%UID/uftrace/$PID.socket. It dies
when the target program finishes.

The libmcount-daemon.so library is chosen when the --daemon option is
used.

Signed-off-by: Clément Guidi <clementguidi@gmail.com>