tui: Rename graph fields and code
To prepare for additional field customization such as report mode,
this commit renames graph fields and code in tui.

If additional modes can commonly use variables or functions
of the graph mode,
then they should be changed to the generic names.
- graph_field_help[] -> field_help[]
- display_graph_field() -> display_tui_field()
- tui_window_graph_field() -> tui_window_field()

If some variables are only for the graph mode but they don't specify it,
then prefix `graph_` is concatenated with the variables to distinguish
from other modes.
- field_total_time -> graph_field_total
  (omit time to reduce name length)
- field_self_time -> graph_field_self
- field_addr -> graph_field_addr

Signed-off-by: Sangwon Hong <qpakzk@gmail.com>