dwarf: Fix two same struct passed by value
It missed to update the reg_idx when the struct is passed via
registers.  Then next struct also passed by registers will see the
previous parameter and think they are duplicate.  The end result was
that it overwrites the previous one and shows just a single result.

Fix it by arg->reg_idx for the last register it has.

  $ uftrace -a -C compare_iters  tests/t-dwarf7
  # DURATION     TID     FUNCTION
              [113893] | main() {
   162.832 us [113893] |   compare_iters(__normal_iterator{...}, __normal_iterator{...}) = 0;
   173.225 us [113893] | } = 0; /* main */

Fixed: #1281

Signed-off-by: Namhyung Kim <namhyung@gmail.com>