event: Fix a crash by adding PROT_READ along with PROT_EXEC
Currently, uftrace gets segfault when patching SDT events into the
target binary.

  $ gcc -pg -o t-sdt s-sdt.c

  $ uftrace -E uftrace:* --match glob t-sdt
  WARN: child terminated by signal: 11: Segmentation fault
  WARN: cannot open record data: /tmp/uftrace-live-CCoVbF: No data available

The reason of the segfault is because the access permission of the
SDT memory region becomes PROT_EXEC only without having PROT_READ.

It might be okay in some environments but some are not so this patch
explicitly adds PROT_READ along with PROT_EXEC.

Before:
  147 event_sdt           : SG SG SG SG SG SG SG SG SG SG
  150 recv_event          : NZ NZ NZ NZ NZ NZ NZ NZ NZ NZ

After:
  147 event_sdt           : OK OK OK OK OK OK OK OK OK OK
  150 recv_event          : NG NG NG NG NG NG NG NG NG NG

The test 150 recv_event is not fully fixed but another problem will be
fixed in the follow-up commit, but this patch fixes the horrible crash
anyway so it looks from NZ to NG now.

Fixed: #1286

Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>