event: Do not ignore when -E option is used with --no-event in replay
In analysis commands such as replay, it currently ignores -E option when
--no-event option is used unlike the behavior in record command.

So this patch makes the analysis commands stop ignoring events that was
recorded with -E option when -E option is explicitly given in analysis
commands again.

Before:
  $ uftrace record --no-event -E uftrace:event t-sdt
  $ uftrace replay --no-event -E uftrace:event
  # DURATION     TID     FUNCTION
     1.552 us [ 14095] | __monstartup();
     0.923 us [ 14095] | __cxa_atexit();
              [ 14095] | main() {
              [ 14095] |   foo() {
    14.415 us [ 14095] |   } /* foo */
    15.687 us [ 14095] | } /* main */

After:
  $ uftrace record --no-event -E uftrace:event t-sdt
  $ uftrace replay --no-event -E uftrace:event
  # DURATION     TID     FUNCTION
     1.552 us [ 14095] | __monstartup();
     0.923 us [ 14095] | __cxa_atexit();
              [ 14095] | main() {
              [ 14095] |   foo() {
              [ 14095] |     /* uftrace:event */
    14.415 us [ 14095] |   } /* foo */
    15.687 us [ 14095] | } /* main */

It has to more precisely detect the given event name with -E, but we
better make it simply stop ignoring when any of event is given with -E
option for now.

And let's use it in the current test failure in 150 recv_event test.

Before:
  150 recv_event          : NG NG NG NG NG NG NG NG NG NG

After:
  150 recv_event          : OK OK OK OK OK OK OK OK OK OK

Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>