build: Fix a build warning with libpython
When there are libpython2.7 and libpython3 together in a system for
some reason, the version.sh script will emit a warning like

  misc/version.sh: 37: test: check-deps/have_libpython2.7: unexpected operator
      ...

It was because the test used a wildcard match and it expanded to two
filenames but the test statement expected only a single expression.
Let's fix it by checking them explicitly.

Fixed: #1297

Signed-off-by: Namhyung Kim <namhyung@gmail.com>