utils: Use only libunwind without using execinfo
The current stacktrace in release mode uses backtrace() in execinfo.h,
but it can't show static function names and it doesn't work reliable in
libmcount*.so libraries during record time.

Moreover, execinfo.h is not well supported in some environments such as
alpine Linux.

So we better use only libunwind when it's available in debug build.
If not, we just do not print anything in stacktrace().

Fixed: #1298

Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>