report: Fix to properly set color by detecting color mode
'sc' means start character and is set to TERM_COLOR_NORMAL initially,
but it has to be set to TERM_COLOR_RESET when color mode is on.

This patch fixes missing reset 'ec' in print_diff_percent and
print_diff_count.

In addition, uftrace report output prints colors even if color has to
be disabled so this provides get_color helper function for this.

Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>