add debian/* from thepaul, updated for 0.10+ and python3
deb packages can now be built on Debian/Ubuntu-based distros
using debuild, sbuild and dpkg-buildpackage.

For this, create an new directroy and in it, git clone this into a
directory named uftrace-0.10.54 (version set in debian/changelog).

For building in debian/unstable (sid) schroot, install and run
sbuild-debian-developer-setup and then run sbuild. See:
https://wiki.debian.org/sbuild