doc: Add doc/run-spell.sh for spell check in the doc
spell is a spell checking program which prints each misspelled word on a
line of its own.

We need to generate dictionary file used in the project first and it was
generated as follows.

  $ spell *.md doc/*.md | sort | uniq --ignore-case > doc/dictionary.txt

The spell checker can be run as follows.

  $ ./doc/run-spell.sh
  /home/honggyu/work/uftrace/doc/uftrace-record.md:633: excutables

It detects an incorrect word from the document.

Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>