Fix VERSION_GIT(version.h) in srcdir from tarball below other git repo
To set the version string to be generate into `version.h`, `Makefile` uses:
```
VERSION_GIT := $(shell git describe --tags 2> /dev/null || echo v$(VERSION))
```
This usually works because ```git describe --tags``` would abort when
$(srcdir) is not located in or below a git repo, and v$(VERSION) is used.

When uftrace is built from a release tarball and the $(srcdir) is located
in a temporary directory below another git worktree or repo, `git describe`
describes the verison of the git repo in which the `$(srcdir)` is located,
which is a completely different version string from the other repo and not
the release version string of the git tarball.

Fix this by checking if $(srcdir)/.git is a directory (or a file):

In case uftrace is checked out externally, e.g. as a git submodule,
.git is a text file with a `git:`, followed by relative path to the
real .git directory, so it works also when .git/ is located externally.

This still does not work in other ways of including one git repo in another
like https://github.com/ingydotnet/git-subrepo, but they are not that common
and I see little reason to do that. I think these cases can only be fixed by
settings the VERSION variable outside of release tags differently as a flag.

Inserting 'test -e .git && ' into the shell command is enough as simple fix.