dynamic: Add a logic to patch functions in __patchable_function_entries
This patch adds an additional logic to patch only the functions found at
__patchable_function_entries section.

It can be cheaper compared to iterate all the available symbols so can
be a bit more efficient.  In addition, this logic can be applied to
striped binaries that don't have symbol information.

The patch_func_matched is divided into two functions as follows.

  static void patch_func_matched(struct mcount_dynamic_info *mdi,
                                struct uftrace_mmap *map)
  {
          if (mdi->type == DYNAMIC_PATCHABLE)
                  patch_patchable_func_matched(mdi, map);
          else
                  patch_normal_func_matched(mdi, map);
  }

The original logic is moved into patch_normal_func_matched() function.

Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>