dynamic: Handle when __patchable_function_entries section is stripped
In some cases, the __patchable_function_entries section can be removed.
For example, -Wl,--gc-sections strips this section.  In this case,
mdi->nr_patch_target becomes 0 so it's not possible to find the
patchable locations from the section.

However, we can still try again by looping over all the symbols
available and check if the function begins with the NOP patterns for
patchable function entry inside patch_normal_func_matched().

This patch also adds t267_patchable_dynamic5.py that tests the binary,
which is compiled with -Wl,--gc-sections option.

  267 patchable_dynamic5  : OK OK OK OK OK OK OK OK OK OK

Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>