dynamic: Set dynamic min_size as patched code size + 1
uftrace has to skip the function target if the function size is same as
the current min_size.  This is needed in the following case.

  00000000000005f5 <foo>:
   5f5:   e9 26 ff ff ff          jmpq   520 <AAA::bar(int)@plt>

  00000000000005fa <AAA::bar(int)>:
   5fa:   48 8b 05 ef 09 20 00    mov    0x2009ef(%rip),%rax        # 200ff0 <_DYNAMIC+0x198>
   601:   89 38                   mov    %edi,(%rax)
   603:   c3                      retq

In this case, the call instruction is overwritten at the entry of 'foo',
but the return address for the entry of foo is misinterpreted as the
address 0x5fa, which is the address of 'AAA::bar(int)>' so uftrace gets
confused about the function name and it fails as follows.

This patch sets min_size as patched code size + 1 and compare if the
function size less than it, then skips patching the function for dynamic
tracing.

Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>