dynamic: Do not skip weak symbol functions for dynamic tracing
If there is a simple example as follows.

  $ cat shared_ptr.cc
  #include <memory>

  int main()
  {
          std::shared_ptr<int> p(new int);
  }

We can compile it with -pg and -fpatchable-function-entry=5 as follows.

  $ g++ -pg -o shared_ptr.pg shared_ptr.cc
  $ g++ -fpatchable-function-entry=5 -o shared_ptr.fpatchable shared_ptr.cc

Tracing result of shared_ptr.fpatchable shows much less functions than
shared_ptr.pg example.

  $ uftrace --no-libcall shared_ptr.pg | wc -l
  39

  $ uftrace -P. --no-libcall shared_ptr.fpatchable | wc -l
  11

It's because many functions are shown as weak symbols.

  $ nm -C shared_ptr.fpatchable | grep "std::shared"
  0000000000400b0e W std::shared_ptr<int>::shared_ptr<int, void>(int*)
  0000000000400b0e W std::shared_ptr<int>::shared_ptr<int, void>(int*)
  0000000000400aee W std::shared_ptr<int>::~shared_ptr()
  0000000000400aee W std::shared_ptr<int>::~shared_ptr()

But our logic in skip_sym() skips such weak symbols so this patch makes
dynamic tracing not to skip such weak symbols.

Fixed: #1384
Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>