filter: Add -L/--loc-filter option to filter by source location
The -L/--loc-flter option supports a filter that traces or does not trace the
selected source location with the debug_info loaded.

These are usage example.

The '--srcline' option is required when using record command.
$ uftrace record --srcline t-lib

The '-L' option is required when using replay command to set a filter to
trace selected source files.
$ uftrace replay -L s-libmain.c --srcline
          [  4636] | main() { /* /home/uftrace/tests/s-libmain.c:16 */
 4.119 us [  4636] |   foo(); /* /home/uftrace/tests/s-libmain.c:11 */
 4.544 us [  4636] | } /* main */

You can set a filter to not trace selected source files with the '@hide' suffix.
$ uftrace replay -L s-libmain.c@hide
          [  5327] | lib_a() {
          [  5327] |   lib_b() {
 1.226 us [  5327] |     lib_c();
 2.419 us [  5327] |   } /* lib_b */
 2.789 us [  5327] | } /* lib_a */

Closes: #1378
Signed-off-by: Eunseon Lee <esintospace@gmail.com>