script: Add event handler support
The 'script_uftrace_event' is to handle event records in the uftrace
data.  The argbuf fields in the script_context will have the event
information in a string format (key=value pairs) if exists.

For example, read/diff-pmu-cache event will have something like below:

  sc_ctx.name   = "diff:pmu-cache"
  sc_ctx.argbuf = "refers=10000 misses=200"

It's the script's responsibility to check and parse the event data
properly.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>