perf: Fix segfault due to invalid cpu index
When setting up the writer_arg for record, it spreads cpu to
parallelize the processing.  But if the number of cpus is not a
multiple of the number of threads, it can have invalid index (-1) at
the end of the array.

It should check the cpu index before accessing the perf mmap.

Fixed: #1418

Signed-off-by: Namhyung Kim <namhyung@gmail.com>