tests: Fix possibly incorrect libmcount.so loading
If the installed libmcount.so is x86_64 based, but the compiled
libmcount.so inside uftrace/libmcount directory is i386 version, then
the test t269 fails because the record command in prerun didn't add
--libmcount-path option so it tries to search libmcount.so in the
default library search path, which has x86_64 version of libmcount.so.

The following error message shows the problem.

  $ ./configure --arch=x86_64
  $ make -j install

  $ ./configure --arch=i386
  $ make -j

  $ ./runtest.py -vdp -O0 269
  Start 1 tests with 1 worker
  Test case                 pg
  ------------------------: O0
  build command: gcc -o t-exp-str -fno-inline -fno-builtin -fno-ipa-cp -m32 -fno-omit-frame-pointer -D_FORTIFY_SOURCE=0  -pg -O0  s-exp-str.c
  ERROR: ld.so: object 'libmcount.so' from LD_PRELOAD cannot be preloaded (wrong ELF class: ELFCLASS64): ignored.

To avoid this, we have to add TestBase.default_opt to record command
option.  The same problem applies to t270 so this is fixed together.

Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>