test: Use info bit mask properly in the data_basic test
It was confused to use the bit index instead of the bit mask.  So
'EXE_NAME | EXIT_STATUS' became 0 | 2 rather than (1 << 0 | 1 << 2).
This is actually the same as (1 << 1) which is BUILD_ID.

In my setup, GCC seems to add a build-id for the binary by default but
clang in the TravisCI seems not.  That makes the test failed - and the
test macro just returned from the function without calling close_data_file().

As a result, we ended up seeing a bunch of memory leaks by Asan.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>