tests: Fix applying zero offset to null pointer in unittest
rb_entry's first argument 'rb_next(&save_entry->node)' is used as
the pointer to the member in container_of.
If it is NULL, it does arithmetic on the null pointer in container_of.
Below is the code.

    (type *)((char *)__mptr - offsetof(type, member));

In this part, offsetof(struct debug_entry, node) is 0. So it outputs
the error as "applying zero offset to null pointer"
This change is to add null checking to fix it.

Fixed: #1441
Signed-off-by: Jia Ha <wendyisdream1004@gmail.com>