dump: Add the option to show function call graph as mermaid diagram
This option shows the function call graph as a mermaid flowchart
for visualization.

- mermaid page : https://mermaid-js.github.io/mermaid/

The usage for this option is add --mermaid in dump command.

The result can be rendered in the browser so it is formatted as HTML.
The mermaid syntax itself can be rendered in other services
including a mermaid plugin like github or draw.io.

Test case 272 in runtest.py is added to verify the mermaid option's result.
The purpose of this test case is to verify to represent the call flow
in mermaid format.
The verbose test example is like below

  ./runtest.py -v 272
  ...
  =========== original ===========
  <html>
  <body>
  <script src="https://cdn.jsdelivr.net/npm/mermaid/dist/mermaid.min.js"></script>
  <script>
    var config = {
    startOnLoad: true,
    maxTextSize: 99999999,
    flowchart: { useMaxWidth: false }
    };
    mermaid.initialize(config);
  </script>
  <h2>Function Call Graph for <span style="color:blue">t-abc</span></h2>
  <div class="mermaid">
  flowchart TB
    0["t-abc"] -->|1| 1["main"];
    1["main"] -->|1| 2["a"];
    2["a"] -->|1| 3["b"];
    3["b"] -->|1| 4["c"];
  </div>
  </body>
  </html>

  ===========  result  ===========
  0_t-abc 1> 1_main
  1_main 1> 2_a
  2_a 1> 3_b
  3_b 1> 4_c
  =========== expected ===========
  0_t-abc 1> 1_main
  1_main 1> 2_a
  2_a 1> 3_b
  3_b 1> 4_c

Signed-off-by: Jia Ha <wendyisdream1004@gmail.com>