record: Do not initialize when libmcount.so is explictly linked
Currently, libmcount.so library cannot be explictly linked to a binary
because it always initializes many things for uftrace record expecting
uftrace.data directoy is already created by uftrace.

Because of this assumption, it shows an error when a binary has an
explicit link to libmcount.so as follows.

  # make sure if there is an existing uftrace.data directory.
  $ rm -rf uftrace.data

  $ gcc -pg -g -o t-hello tests/s-hello.c -L./libmcount -lmcount
  $ ldd t-hello | grep libmcount
          libmcount.so => /home/honggyu/usr/lib/libmcount.so (0x00007fad4b99e000)

  $ ./t-hello
  mcount: /home/honggyu/work/uftrace/libmcount/record.c:1127:record_proc_maps
   ERROR: cannot open for writing maps file: No such file or directory

This patch is to avoid this error by adding UFTRACE_TRACING_ON env
variable to check whether libmcount.so init is invoked with uftrace or
not.  Having this patch, it doesn't show an error and the target binary
can still be traced by uftrace.

  $ ./t-hello
  Hello world

  $ uftrace ./t-hello
  Hello world
  # DURATION     TID     FUNCTION
              [ 40432] | main() {
    13.835 us [ 40432] |   printf();
    17.606 us [ 40432] | } /* main */

If debug mode is on then it shows that libmcount.so is not preloaded.

  $ uftrace record -v t-hello
      ...
  uftrace: libmcount.so is already linked so do not preload it.
      ...
  Hello world

Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>