tests: Fix random failures in diff tests
The current diff tests have a common tricky problem because we can't
guarantee the difference of execution time for two different executions.

Because of this problem, some diff tests randomly fails as follows.

  158 report_diff_policy1 : OK OK OK NG OK OK NG OK OK NG  OK OK OK OK OK OK OK OK OK OK
  159 report_diff_policy2 : OK OK OK OK OK OK OK OK OK OK  OK OK OK NG OK OK OK OK OK OK
  160 report_diff_policy3 : NG OK OK OK OK OK OK NG NG OK  OK OK OK OK OK OK OK OK OK OK
  240 report_diff_field2  : OK OK OK NG OK OK NG NG OK OK  NG OK OK OK NG OK OK NG NG OK
  241 report_diff_field3  : OK OK OK OK OK OK OK OK OK OK  OK OK OK OK OK OK OK OK NG OK
  242 report_diff_field4  : OK OK OK OK OK OK OK OK OK OK  OK NG OK OK OK OK OK OK OK OK
  243 report_diff_field5  : OK OK OK OK OK OK OK OK OK NG  OK NG OK OK OK OK OK OK OK OK

I have reproduced the problem in 158 and the report diff results shows
as follows in both pass and fail cases.

The pass case:
  $ uftrace report -d xxx/ --diff yyy
        ...
      Total time     Self time         Calls   Function
     ===========   ===========   ===========   ====================
       +1.348 ms     +0.147 us            +0   main
       +1.344 ms     +1.344 ms            +3   usleep
       +1.277 ms     +2.501 us            +0   foo
     +161.926 us     +1.410 us            +0   bar
       +0.022 us     +0.022 us            +0   atoi

The fail case:
  $ uftrace report -d xxx/ --diff yyy
        ...
      Total time     Self time         Calls   Function
     ===========   ===========   ===========   ====================
       +1.275 ms     +1.275 ms            +3   usleep
       +1.274 ms     -0.491 us            +0   main
       +1.208 ms     +0.031 us            +0   foo
     +153.580 us     +0.168 us            +0   bar
       -0.496 us     -0.496 us            +0   atoi

The fail case shows 'usleep' at the top and that means the total time
diff 'usleep' is greater than 'main', which is a different result.

Since we can't guarantee that the execution difference of two different
records, we should use a reliable field for sorting the result.

The same problem happens in many diff tests, so this patch fixes all
those problem.

It also changes s-diff.c source to make the test result more clearly
distinguished.

The following report test also shows the same problem so it's fixed
together.

  025 report_s_call       : OK OK OK OK OK OK OK OK NG OK  OK OK OK OK OK OK OK OK OK OK

Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>