uftrace: Do not override time range
When uftrace record was called with a time filter, it saves the time
filter in the default.opts file to apply it for kernel functions and
events too when replay.

But if user requested a time range during the replay, the time filter
in the default option overrides the range so users cannot use the -r
option.  Let's ignore the default time filter when user specified a
time range in the command line.

Fixed: #1593

Signed-off-by: Namhyung Kim <namhyung@gmail.com>