uftrace: Add a simple benchmark
I added a simple program to benchmark uftrace record performance.
The bench program calls foo(), bar() and baz() functions repeatedly
(by default a half of 1 million times, but you can change it with
arguments).  The foo() is a leaf function so it's gonna be the base
for the performance comparison.  And the bar() calls baz() and we can
use them it to get the precise performance numbers.

Here's the pseudo code for the benchmark functions:
   foo():                        <libmcount.so>
A    // prologue
B    // calls mcount
                                mcount():
C                                  // save registers/arguments
D                                  // check filters/triggers
E                                  // setup rstack
F                                  // save timestamp       <-----------+
G                                  // hook return address              |
H                                  // restore registers/arguments      |
I                                  // go back to the original function |
J    // do the function body                                           | = U_a
K    // return                                                         |
                                mcount_exit()                          |
L                                  // do some minimal checking         |
M                                  // save timestamp       <-----------+
N                                  // check filters/triggers
O                                  // save records to shmem
P                                  // return to the original caller

Note that uftrace can only record time from step F to step M for the
function foo(). For the steps C, D, E and N, O, P, we need to use nested
functions like bar() and baz().

For simplicity, let's call the portion of overhead in uftrace that's
covered by the leaf function (F-I and L-M) as U_a, and rest of them as
U_b (for C-E and N-P).  Also let's say the overhead of the original
function as O.  So the performance of each functions should be:

  foo   = O_foo + U_a + U_b
                        ^^^ (not measured)
  ${foo_self} = O_foo + U_a

We want to measure the performance of 'foo' but what uftrace shows is
just actually 'O_foo + U_a' and we don't know U_b.  We know O_foo is
very small as it's a leaf function with a memory operation for the same
address (it should hit the L1 cache most likely).

To determine the U_b part, we can use the nested functions:

  bar   = O_bar + U_a + U_b + ${baz}
    baz = O_baz + U_a + U_b

But in this case, what uftrace shows for bar(total) is actually:

  ${bar_total} = O_bar + U_a + (O_baz + U_a + U_b)
                                -----------
                                ${baz_self}
  O_bar ~= O_baz

  ${bar_self}  = ${bar_total} - ${baz_self}
               = O_bar + U_a + U_b         // actual overhead
              ~= O_baz + U_a + U_b
               = ${baz_self} + U_b

  U_b = ${bar_self} - ${baz_self}

Now we can get the U_b part, but it still has some issue.  Because the
step O (save rstacks to shmem) contains some records for the parent
functions as uftrace postpones the recording of entry functions to the
exit functions.  So step O in the baz function will record 3 data:

  1 entry of bar + 1 entry of baz + 1 exit of baz

But we cannot know exact overhead of the step O so we will ignore the
issue for now.  So the self-time of bar() already contains U_b part
and eliminated self-time of baz().  Then it's the precise overhead of
a single function for uftrace.

The bench.sh script is to run the benchmark easily.  It comes with a
couple of command line options to setup the benchmark parameter.

  $ misc/bench.sh -h
  Usage: bench.sh [OPTION]
    OPTION  -c N      Use CPU N during the benchmark. (default: 3)
            -p PROG   Use program PROG. (default: ./bench)
            -u UOPT   Use uftrace option UOPT. Please quote it.
            -v        Show verbose messages.
            -h        Show this help and exit.

It runs the benchmark program on the dedicated CPU and set the cpufreq
for the full performance.  So it needs to root privilege.

A new make target 'bench' was added for even simpler usage like:

  $ make bench
    CC       uftrace.o
    LINK     uftrace

  # uftrace bench
      Self avg    Self min  Function
    ==========  ==========  ====================
      0.112 us    0.070 us  bar            <---- actual overhead
      0.029 us    0.025 us  baz
      0.029 us    0.026 us  foo

You can pass some options using BENCHARG variable

  $ make bench BENCHARG='-v -c 1'

Signed-off-by: Namhyung Kim <namhyung@gmail.com>