test: Fix lib name test on Debian
The lib name test (192) checks if it prints the library names properly.
The getpid() is from libc but its name can be different on each distro.
Usually it shows the version number like "libc-2.35.so" but for some
reason my system shows the generic SONAME like "libc.so.6".

Change the default (expected) result to have the soname as we can try
the versioned name in the fixup routine.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>